+++ Pandora/Pattern/Functor/Monad.hs
@@ -2,7 +2,8 @@ module Pandora.Pattern.Functor.Monad where
 
 import Pandora.Pattern.Functor.Covariant (Covariant)
 import Pandora.Pattern.Functor.Bindable (Bindable)
-import Pandora.Pattern.Functor.Pointable (Pointable)
+import Pandora.Pattern.Functor.Monoidal (Monoidal)
+import Pandora.Paradigm.Primary.Algebraic.Product ((:*:))
 
 {- |
 > Let f :: (Pointable t, Bindable t) => a -> t a
@@ -18,7 +19,7 @@ import Pandora.Pattern.Functor.Pointable (Pointable)
 --infixl 1 >>=-, ->>=
 --infixr 1 -=<<, =<<-
 
-class (Covariant t (->) (->), Pointable t (->), Bindable t (->)) => Monad t where
+class (Covariant t (->) (->), Monoidal t (->) (->) (:*:) (:*:), Bindable t (->)) => Monad t where
 	--(>>=-) :: t a -> t b -> t a
 	--(>>=-) x y = x >>= \r -> y >>= \_ -> point r
 	--(->>=) :: t a -> t b -> t b